Fix mean count computation and comment

+++ SolderStation/solder_station.cpp
@@ -88,7 +88,7 @@ void set_iron_temperature(int tTmp) {
 #if TEMP_MEAN < DELAY_MAIN_LOOP*2
 #error TEMP_MEAN should be greater than DELAY_MAIN_LOOP*2
 #endif
-#define MEAN_COUNT ((float) (TEMP_MEAN/(DELAY_MAIN_LOOP*2))) // Only one period on two are used for measurement
+#define MEAN_COUNT (((float)TEMP_MEAN)/((float)(DELAY_MAIN_LOOP*2))) // *2 because we have only one period on two is used for measurement
   float temp_iron_temperature = iron_temperature * (MEAN_COUNT - 1.0f);
   temp_iron_temperature += ((float)tTmp);
   iron_temperature = temp_iron_temperature / MEAN_COUNT;
